% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{datajud_count}
\alias{datajud_count}
\title{Count total results for a DataJud query}
\usage{
datajud_count(cfg, body)
}
\arguments{
\item{cfg}{List. Configuration object from \code{\link[=datajud_config]{datajud_config()}}.}

\item{body}{List. Elasticsearch query body from \code{\link[=build_query]{build_query()}}.}
}
\value{
Integer. Total number of matching documents.
}
\description{
Count total results for a DataJud query
}
