% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{datajud_estimate_download}
\alias{datajud_estimate_download}
\title{Estimate download time for a DataJud query}
\usage{
datajud_estimate_download(cfg, body, page_size = 100)
}
\arguments{
\item{cfg}{List. Configuration object from \code{\link[=datajud_config]{datajud_config()}}.}

\item{body}{List. Elasticsearch query body from \code{\link[=build_query]{build_query()}}.}

\item{page_size}{Integer. Results per page. Default \code{100}.}
}
\value{
An object of class \code{datajud_estimate}.
}
\description{
Estimate download time for a DataJud query
}
