% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{print.datajud_estimate}
\alias{print.datajud_estimate}
\title{Print method for datajud_estimate}
\usage{
\method{print}{datajud_estimate}(x, ...)
}
\arguments{
\item{x}{A \code{datajud_estimate} object.}

\item{...}{Ignored.}
}
\description{
Print method for datajud_estimate
}
