% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints_processos.R
\name{processos}
\alias{processos}
\title{Retrieve all processes from a tribunal}
\usage{
processos(cfg, ...)
}
\arguments{
\item{cfg}{List. Configuration object from \code{\link[=datajud_config]{datajud_config()}}.}

\item{...}{Additional arguments passed to \code{\link[=datajud_search]{datajud_search()}}.}
}
\value{
A tibble with all processes.
}
\description{
Retrieve all processes from a tribunal
}
