% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders.R
\name{query_match_all}
\alias{query_match_all}
\title{Build an Elasticsearch match_all query}
\usage{
query_match_all()
}
\value{
A named list representing the query clause.
}
\description{
Build an Elasticsearch match_all query
}
