% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders.R
\name{query_term}
\alias{query_term}
\title{Build an Elasticsearch term query}
\usage{
query_term(field, value)
}
\arguments{
\item{field}{Character. Field name.}

\item{value}{Character. Exact value to match.}
}
\value{
A named list representing the query clause.
}
\description{
Build an Elasticsearch term query
}
