% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints_processos.R
\name{processo_por_numero}
\alias{processo_por_numero}
\title{Search a process by its number}
\usage{
processo_por_numero(cfg, numero_processo, ...)
}
\arguments{
\item{cfg}{List. Configuration object from \code{\link[=datajud_config]{datajud_config()}}.}

\item{numero_processo}{Character. Full process number.}

\item{...}{Additional arguments passed to \code{\link[=datajud_search]{datajud_search()}}.}
}
\value{
A tibble with matching processes.
}
\description{
Search a process by its number
}
